pV2, UTKFace, stats outputted, confusion matrices.
ProjectV2, uses UTKFace, statistics summary outputted in separate file, confusion matrix for age and gender outputted in separate files. Each image has the potential to be enhanced depending on its quality (very little of the time does the enhancement help). Each image tested shows actual versus predicted, if it was enhanced (and if it helped), OR if no faces / multiple faces were detected.